Move project component files to worklists folder
