--- src/app/projects/data-types/task.ts
@@ -1,9 +1,0 @@
-export class Task {
-    id: number;
-    start_date: string;
-    text: string;
-    progress: number;
-    duration: number;
-    parent: number;
-    type: string;
-}
