--- src/app/projects/projects-routing.module.ts
@@ -1,28 +1,0 @@
-/**
- * @license
- * Copyright (c) 2017 La Vía Óntica SC, Ontica LLC and contributors. All rights reserved.
- *
- * See LICENSE.txt in the project root for complete license information.
- *
- */
-
-import { NgModule } from '@angular/core';
-import { RouterModule } from '@angular/router';
-
-import { SecurityGuardService } from '../core';
-
-import { MainLayoutComponent } from '../shared';
-
-import { ProjectsMainPageComponent } from '../projects/main-page/projects-main-page.component';
-
-@NgModule({
-  imports: [
-    RouterModule.forChild([
-      {
-        path: 'projects', component: MainLayoutComponent, canActivate: [SecurityGuardService],
-        children: [{ path: 'search', component: ProjectsMainPageComponent }]
-      }
-    ])],
-  exports: [RouterModule]
-})
-export class ProjectsRoutingModule { }
