+++ src/app/worklists/create-activity/activity-editor.component.scss
