+++ src/app/worklists/data-types/task.ts
@@ -40,3 +40,14 @@ export interface TaskRef {
 
     return empty;
   }
+
+  //**** */
+  export class Task {
+    id: number;
+    start_date: string;
+    text: string;
+    progress: number;
+    duration: number;
+    parent: number;
+    type: string;
+}
