+++ src/app/worklists/editor/task-form-editor.component.html
@@ -61,7 +61,7 @@
       <tr>
         <td>Progreso:</td>
         <td>
-          <select class="select-box" style="width:60px">
+          <select class="select-box" style="width:60px" [(ngModel)]="selectedTask.progress"> 
           <option value="0">0%</option>
           <option value="10">10%</option>
           <option value="20">20%</option>
