+++ src/app/worklists/gantt/viewer.component.scss
